+++ src/async/handlers/tsonPromise.ts
@@ -1,5 +1,5 @@
-import { TsonAsyncType } from "../async/asyncTypes.js";
-import { TsonPromiseRejectionError } from "../errors.js";
+import { TsonPromiseRejectionError } from "../../errors.js";
+import { TsonAsyncType } from "../asyncTypes.js";
 
 function isPromise(value: unknown): value is Promise<unknown> {
 	return (
